[Comgr][hotswap] Share the AMDGPU registration across the test binaries
The test binaries link the hotswap OBJECT libraries without linking amd_comgr,
so each one has to supply COMGR::ensureLLVMInitialized itself. amd_comgr bakes
in its own copy of LLVM and hides every internal symbol, so linking it for that
one function would register AMDGPU into that copy's registry while the binary's
own LLVM kept an empty one. Eleven binaries need the definition and eleven
carried their own, identical down to the order of the initializer calls.

State it once instead, in a translation unit compiled into each of them. It sits
under test-support rather than in either test directory because the consumers
span both: the transpile driver under test-lit and the hotswap unit tests under
test-unit. An OBJECT library rather than a source listed per target, so the
registration is compiled once and the same object handed to each of them.
