+++ amd/comgr/CMakeLists.txt
@@ -696,6 +696,7 @@ if(BUILD_TESTING)
   if (NOT CMAKE_SOURCE_DIR STREQUAL CMAKE_CURRENT_SOURCE_DIR)
     set_property(GLOBAL APPEND PROPERTY LLVM_ADDITIONAL_TEST_TARGETS check-comgr)
   endif()
+  add_subdirectory(test-support)
   add_subdirectory(test)
   add_subdirectory(test-lit)
   add_subdirectory(test-unit)
