+++ amd/comgr/test-unit/hotswap/raiser/UserSgprLayoutTest.cpp
@@ -20,29 +20,9 @@
 
 #include "llvm/Support/AMDHSAKernelDescriptor.h"
 #include "llvm/Support/Error.h"
-#include "llvm/Support/TargetSelect.h"
 #include "llvm/Support/raw_ostream.h"
 
 #include "gtest/gtest.h"
-
-#include <mutex>
-
-// initMCState calls COMGR::ensureLLVMInitialized, whose production definition
-// lives in libamd_comgr. Provide the registration here so the test binary stays
-// minimal instead of linking the full Comgr.
-namespace COMGR {
-void ensureLLVMInitialized() {
-  static std::once_flag Once;
-  std::call_once(Once, [] {
-    LLVMInitializeAMDGPUTargetInfo();
-    LLVMInitializeAMDGPUTargetMC();
-    LLVMInitializeAMDGPUDisassembler();
-    LLVMInitializeAMDGPUAsmParser();
-    LLVMInitializeAMDGPUAsmPrinter();
-    LLVMInitializeAMDGPUTarget();
-  });
-}
-} // namespace COMGR
 
 using namespace COMGR::hotswap;
 using namespace llvm;
