+++ amd/comgr/test-unit/hotswap/raiser/WaveProjectionTest.cpp
@@ -32,29 +32,8 @@
 #include "llvm/IR/LLVMContext.h"
 #include "llvm/IR/Module.h"
 #include "llvm/IR/Type.h"
-#include "llvm/Support/TargetSelect.h"
 
 #include "gtest/gtest.h"
-
-#include <mutex>
-
-// The wave-projection objects link the hotswap::decoder MC stack, whose
-// initMCState calls COMGR::ensureLLVMInitialized; its production definition
-// lives in libamd_comgr. Provide the registration here so the test binary stays
-// minimal instead of linking the full Comgr.
-namespace COMGR {
-void ensureLLVMInitialized() {
-  static std::once_flag Once;
-  std::call_once(Once, [] {
-    LLVMInitializeAMDGPUTargetInfo();
-    LLVMInitializeAMDGPUTargetMC();
-    LLVMInitializeAMDGPUDisassembler();
-    LLVMInitializeAMDGPUAsmParser();
-    LLVMInitializeAMDGPUAsmPrinter();
-    LLVMInitializeAMDGPUTarget();
-  });
-}
-} // namespace COMGR
 
 using namespace llvm;
 using COMGR::hotswap::initMCState;
